Create fixture for test server creation
